Add a character's custom rolls to the menu. Their invocation does not function yet.
This change requires a corresponding change in the Sentiment system code to populate the system's RollType enum into CONFIG to be read by this module.

Additionally, because a character's custom rolls don't have unique identifiers, only (non-unique) names, it's not possible to reliably map an action to the invocation of the correct custom roll. Achieving that will require assigning unique IDs to custom rolls in the Sentiment system module, which probably means promoting them to Documents instead of just nested JSON objects within the character schema. That's a more involved change, so I'm cutting off this commit here while I decide how to proceed.