Remove requiredCoreModuleVersion property from api object.
Token Action HUD Core has updated its version in newer releases that are compatible with new releases of Foundry VTT. Testing indicates that Token Action HUD Sentiment is compatible with both the older version of Core it was developed against (1.5.6) and the newest version (2.0.16.) However, Core has no semantic for allowing a system module to specify a minimum version, or multiple allowed versions, which would be the ideal solution. Passing a value for requiredCoreModuleVersion will require us to be recognized as compatible by either the old Core module or the new one, but not both, even though we are compatible with both in practice. Therefore, I'm removing the specification altogether, even though that does potentially make us fail later if a user tries to pair us with a really old Core module or if there is an even newer release of Core with which we really aren't compatible. It's not ideal, but this is the best I can do without a more robust versioning parameter to pass to Core.